refactor(addons): Allow sensor execution without native addons
Native addon installation fails when compilers are missing or when executed on unsupported system.
Nevertheless, we still want to support these systems. We can mark these dependency containing addons
as optional and have metric implementations that respect the optional nature to improve platorm
support. In line with current best practices we also define a flag which can be inspected to
determine whether this monitored app has libuv and gc monitoring support.