feat: do not instrument npm or yarn
If a process is started with `npm start` or `yarn start`, two separate
Node.js processes will be started. The first is `node /path/to/npm`
(or `node /path/to/yarn`). This process will kick off another process
with the actual application (for example, `node /usr/src/app`).

With this change, even when NODE_OPTIONS is set to
"--require /usr/src/app/node_modules/@instana/collector/src/immediate",
only the actual application will be instrumented, but not the npm/yarn
process.