fix(serverless): force agent-base@6.0.2 version
https-proxy-agent specifies a loose version range for its dependency
agent-base ("agent-base": "6"). In Node.js 15 and 16, agent-base@6.0.0
is broken for the use case of an HTTPS connection with an HTTP proxy.
Thus, we need to ensure that agent-base@6.0.2 is used, which contains
the fix
(https://github.com/TooTallNate/node-agent-base/pull/47/commits/a5c5c831fdb861dfef498f053740edfed134ee9b).

To do that, we include agent-base@6.0.2 explicitly in
packages/serverless/package.json. If https-proxy-agent releases an
update with a version range for agent-base that includes this fix, we
can remove that dependency again (we do not actually depend on
agent-base directly).