chore(deps): remove node-gyp dependency from @instana/shared-metrics
The main reason behind this change is the increased number of requests
from people using dependency vulnerability scanners (Snyk, Dependabot
etc.). The package node-gyp brings in a lot of dependencies. _None_ of
the reports we have received so far would have been actually relevant
security issues. Usually the story goes like this:
* A vulnerability is reported against some utility package. The
vulnerability would be exploitable if the package is used in a user
facing service _and_ the service would need to pass unsanitized user
input to that utility package.
* The vulnerability report is picked up by the dependency scanners.
* A lot of project use the utility package in completely different
contexts. Most of the times, no user input is anywhere near the usage.
All the dependencies of node-gyp are exclusively used to install
packages from the registry etc.
* Still, the dependency checkers tools report these vulnerabilities for
all projects that use the utility package somewhere, somehow.
* The people using those tools usually do not bother to analyze the
actual usage of the "vulnerable" package, instead, they just submit
ticktes/change requests to their direct downstream dependencies, asking
them to "fix" the vulnerability. This creates tons of useless busy work,
and is generally a huge waste of time for the whole industry.

TL;DR: This removes node-gyp to decrease the likelihood of receiving
these pointless requests in the future.

The dependency from @instana/autoprofile to node-gyp will be removed in
a separate follow-up commit.