test(tracing): fix nats tests after updating to nats@1.4.12
Due to some internal changes in the nats@1.x package in version 1.4.12, the
behaviour is slightly different, in particular with regard to how errors are
handled. Publisher errors would have been thrown synchronously before, but are
now emitted as an event. This required some changes in the integration test
for nats.