refactor(test): remove duplicate method, add eslint rule
Commit a3b3fbf57 accidentally duplicated the method
AgentStubControls#waitUntilAppIsCompletelyInitialized. The first
definition became dead code at that point, because the second definition
would override the first. Later commits then changed the second definition
of the method.

This removes the first definition and moves the effective definition of
the method up.

This also adds a new eslint rule to prevent the same mistake in the
future.