chore: updated lerna.json
We had the situation that `lerna publish` didn't want to publish all packages.
At this time we only had changes in the `instana/aws-lambda` package.

The command we've used was:
NPM_CONFIG_OTP=xxxx lerna publish --force-publish=* && lerna bootstrap

Therefor we want to add this option to our lerna.json to ensure we never run into this again.