test: work around SIGTERM handling conflict between nyc and Prisma
Both nyc and Prisma do fancy things with signal handling. For that
reason, the Prisma test app ignores the SIGTERM signal when the tests
are run via nyc.

This commit works around this issue by sending a SIGKILL one second
later, when SIGTERM does not terminate the child process (the
application under test).